+++ option_supersede.go
@@ -8,18 +8,12 @@ func SupersededBy(s VariableSet, options ...SupersededByOption) DeprecatedOption
 	return option{
 		ApplyToSpecInDeprecatedSet: func(b variable.SpecBuilder) {
 			for _, v := range s.variables() {
-				rel := variable.Supersedes{
-					Subject:    v.Spec(),
-					Supersedes: b.Peek(),
-				}
-
-				for _, opt := range options {
-					opt.applySupersedesOption(&rel)
-				}
-
-				if err := variable.AddRelationship(rel); err != nil {
-					panic(err.Error())
-				}
+				variable.EstablishRelationships(
+					variable.Supersedes{
+						Subject:    v.Spec(),
+						Supersedes: b.Peek(),
+					},
+				)
 			}
 		},
 	}
@@ -27,5 +21,5 @@ func SupersededBy(s VariableSet, options ...SupersededByOption) DeprecatedOption
 
 // SupersededByOption changes the behavior of the SupersededBy() option.
 type SupersededByOption interface {
-	applySupersedesOption(*variable.Supersedes)
+	future()
 }