+++ option_test.go
@@ -1,10 +1,13 @@
 package ferrite_test
 
 import (
+	"fmt"
+	"os"
+
 	"github.com/dogmatiq/ferrite"
 )
 
-func Example_seeAlso() {
+func ExampleSeeAlso() {
 	defer example()()
 
 	verbose := ferrite.
@@ -20,7 +23,7 @@ func Example_seeAlso() {
 	// Output:
 }
 
-func Example_supersededBy() {
+func ExampleSupersededBy() {
 	defer example()()
 
 	verbose := ferrite.
@@ -34,4 +37,72 @@ func Example_supersededBy() {
 	ferrite.Init()
 
 	// Output:
+}
+
+func ExampleRelevantIf_whenRelevant() {
+	defer example()()
+
+	widgetEnabled := ferrite.
+		Bool("FERRITE_WIDGET_ENABLED", "enable the widget").
+		Required()
+
+	widgetSpeed := ferrite.
+		Unsigned[uint]("FERRITE_WIDGET_SPEED", "set the speed of the widget").
+		Optional(ferrite.RelevantIf(widgetEnabled))
+
+	os.Setenv("FERRITE_WIDGET_SPEED", "100")
+	os.Setenv("FERRITE_WIDGET_ENABLED", "true")
+	ferrite.Init()
+
+	if x, ok := widgetSpeed.Value(); ok {
+		fmt.Println("value is", x)
+	} else {
+		fmt.Println("value is not relevant")
+	}
+
+	// Output:
+	// value is 100
+}
+
+func ExampleRelevantIf_whenNotRelevant() {
+	defer example()()
+
+	widgetEnabled := ferrite.
+		Bool("FERRITE_WIDGET_ENABLED", "enable the widget").
+		Required()
+
+	ferrite.
+		Unsigned[uint]("FERRITE_WIDGET_SPEED", "set the speed of the widget").
+		Required(ferrite.RelevantIf(widgetEnabled))
+
+	// FERRITE_WIDGET_SPEED is "required" but we can leave it undefined when
+	// FERRITE_WIDGET_ENABLED is "false" without causing a validation failure.
+	os.Setenv("FERRITE_WIDGET_ENABLED", "false")
+	ferrite.Init()
+
+	// Output:
+}
+
+func ExampleRelevantIf_whenNotRelevantBuInvalid() {
+	defer example()()
+
+	widgetEnabled := ferrite.
+		Bool("FERRITE_WIDGET_ENABLED", "enable the widget").
+		Required()
+
+	ferrite.
+		Unsigned[uint]("FERRITE_WIDGET_SPEED", "set the speed of the widget").
+		Required(ferrite.RelevantIf(widgetEnabled))
+
+	// FERRITE_WIDGET_SPEED is not required because FERRITE_WIDGET_ENABLED is
+	// "false". We want to see the error message but not terminate execution.
+	os.Setenv("FERRITE_WIDGET_SPEED", "-100")
+	os.Setenv("FERRITE_WIDGET_ENABLED", "false")
+	ferrite.Init()
+
+	// Output:
+	// Environment Variables:
+	//
+	//    FERRITE_WIDGET_ENABLED  enable the widget              true | false    ✓ set to false
+	//  ❯ FERRITE_WIDGET_SPEED    set the speed of the widget    <uint>          ✗ set to -100, expected integer
 }